pyproject.toml
[project]
name = "banke"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "ipykernel>=7.0.1",
  "ipywidgets>=8.1.7",
  "matplotlib>=3.10.7",
  "nats-py>=2.11.0",
  "networkx>=3.5",
  "pandas>=2.3.3",
  "scikit-learn>=1.7.2",
  "scipy>=1.16.2",
  "seaborn>=0.13.2",
  "torch-geometric>=2.7.0",
  "torch>=2.9.0",
  "umap-learn>=0.5.9.post2",
]


[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
